% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fm_utils.R
\name{as_p2p}
\alias{as_p2p}
\title{Convert Functional Map to Pointwise Map}
\usage{
as_p2p(fit, reverse = FALSE, use_adj = FALSE)
}
\arguments{
\item{fit}{`fm_fit` object.}

\item{reverse}{If `TRUE`, return source-to-target mapping.}

\item{use_adj}{If `TRUE`, use adjoint-style embeddings.}
}
\value{
Integer index vector representing nearest-neighbor pointwise map.
}
\description{
Convert Functional Map to Pointwise Map
}
