% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fm_utils.R
\name{fm_compose}
\alias{fm_compose}
\title{Compose Two Functional Fits}
\usage{
fm_compose(fit_ab, fit_bc)
}
\arguments{
\item{fit_ab}{`fm_fit` from domain A to B.}

\item{fit_bc}{`fm_fit` from domain B to C.}
}
\value{
`fm_fit` object mapping A to C.
}
\description{
Compose Two Functional Fits
}
