% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fm_consensus.R
\name{fm_consensus_load}
\alias{fm_consensus_load}
\title{Load Consensus Object}
\usage{
fm_consensus_load(path)
}
\arguments{
\item{path}{Path created by [fm_consensus_save()].}
}
\value{
`fm_consensus` object.
}
\description{
Load Consensus Object
}
