% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fm_consensus.R
\name{fm_consensus_save}
\alias{fm_consensus_save}
\title{Save Consensus Object}
\usage{
fm_consensus_save(object, path)
}
\arguments{
\item{object}{`fm_consensus` object.}

\item{path}{Output file path.}
}
\value{
Input `path`, invisibly.
}
\description{
Save Consensus Object
}
