% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fm_domain.R
\name{fm_domain_generic}
\alias{fm_domain_generic}
\title{Generic Domain Adapter}
\usage{
fm_domain_generic(n_samples, data = NULL, ...)
}
\arguments{
\item{n_samples}{Number of samples in the domain.}

\item{data}{Optional payload.}

\item{...}{Passed to [fm_new_domain()].}
}
\value{
An `fm_domain_generic` object.
}
\description{
Generic Domain Adapter
}
