% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fm_domain.R
\name{fm_domain_graph}
\alias{fm_domain_graph}
\title{Graph Domain Adapter}
\usage{
fm_domain_graph(adjacency, ...)
}
\arguments{
\item{adjacency}{Adjacency matrix.}

\item{...}{Passed to [fm_new_domain()].}
}
\value{
An `fm_domain_graph` object.
}
\description{
Graph Domain Adapter
}
