% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fm_domain.R
\name{fm_domain_mesh}
\alias{fm_domain_mesh}
\title{Mesh Domain Adapter}
\usage{
fm_domain_mesh(vertices, faces = NULL, ...)
}
\arguments{
\item{vertices}{Vertex matrix.}

\item{faces}{Optional face index matrix.}

\item{...}{Passed to [fm_new_domain()].}
}
\value{
An `fm_domain_mesh` object.
}
\description{
Mesh Domain Adapter
}
