% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fm_domain.R
\name{fm_domain_pointcloud}
\alias{fm_domain_pointcloud}
\title{Point Cloud Domain Adapter}
\usage{
fm_domain_pointcloud(points, ...)
}
\arguments{
\item{points}{Point matrix.}

\item{...}{Passed to [fm_new_domain()].}
}
\value{
An `fm_domain_pointcloud` object.
}
\description{
Point Cloud Domain Adapter
}
