% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fm_mesh.R
\name{fm_example_mesh_pair}
\alias{fm_example_mesh_pair}
\title{Load a Bundled Example Mesh Pair}
\usage{
fm_example_mesh_pair(
  name = c("posed_tube", "tosca_cat1", "tosca_cat2", "tosca_cat6", "tosca_cat10")
)
}
\arguments{
\item{name}{Example mesh-pair name. Available options are `"posed_tube"`,
`"tosca_cat1"`, `"tosca_cat2"`, `"tosca_cat6"`, and `"tosca_cat10"`.}
}
\value{
List containing `source`, `target`, `truth`, and `metadata`.
}
\description{
Load a Bundled Example Mesh Pair
}
