% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fm_match.R
\name{fm_fit_matrix}
\alias{fm_fit_matrix}
\title{Extract Functional Map Matrix}
\usage{
fm_fit_matrix(fit)
}
\arguments{
\item{fit}{`fm_fit` object.}
}
\value{
Functional map matrix.
}
\description{
Extract Functional Map Matrix
}
