% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fm_utils.R
\name{fm_inverse}
\alias{fm_inverse}
\title{Invert a Functional Fit}
\usage{
fm_inverse(fit, method = c("pseudoinverse", "transpose"))
}
\arguments{
\item{fit}{`fm_fit` object.}

\item{method}{One of `"pseudoinverse"` or `"transpose"`.}
}
\value{
`fm_fit` object with swapped source/target direction.
}
\description{
Invert a Functional Fit
}
