% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fm_domain.R
\name{fm_measure_matrix}
\alias{fm_measure_matrix}
\title{Build a Measure Matrix from Domain Measure}
\usage{
fm_measure_matrix(domain)
}
\arguments{
\item{domain}{`fm_domain` object.}
}
\value{
Dense or sparse measure matrix.
}
\description{
Build a Measure Matrix from Domain Measure
}
