% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fm_domain.R
\name{fm_n_samples}
\alias{fm_n_samples}
\title{Number of Samples in a Domain}
\usage{
fm_n_samples(domain)
}
\arguments{
\item{domain}{`fm_domain` object.}
}
\value{
Integer sample count.
}
\description{
Number of Samples in a Domain
}
