% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fm_network.R
\name{fm_network}
\alias{fm_network}
\title{Create a Functional-Map Network}
\usage{
fm_network(domains, maps = NULL, directed = TRUE)
}
\arguments{
\item{domains}{Named list of `fm_domain` objects.}

\item{maps}{Optional named list of maps keyed as `"i->j"`.}

\item{directed}{Whether to treat edges as directed.}
}
\value{
`fm_network` object.
}
\description{
Create a Functional-Map Network
}
