% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fm_network.R
\name{fm_network_add_map}
\alias{fm_network_add_map}
\title{Add a Map to an FM Network}
\usage{
fm_network_add_map(
  network,
  i,
  j,
  map,
  weight = 1,
  add_reverse = !network$directed
)
}
\arguments{
\item{network}{`fm_network` object.}

\item{i}{Source domain name.}

\item{j}{Target domain name.}

\item{map}{`fm_fit` object or matrix.}

\item{weight}{Edge weight.}

\item{add_reverse}{For undirected networks, add reverse map automatically.}
}
\value{
Updated `fm_network` object.
}
\description{
Add a Map to an FM Network
}
