% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fm_diagnostics.R
\name{fm_network_compare}
\alias{fm_network_compare}
\title{Compare Pre/Post Network Diagnostics}
\usage{
fm_network_compare(pre, post)
}
\arguments{
\item{pre}{Pre-sync `fm_network` object.}

\item{post}{Post-sync `fm_network` object.}
}
\value{
`fm_network_report` object.
}
\description{
Compare Pre/Post Network Diagnostics
}
