% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fm_network.R
\name{fm_network_cycle_error}
\alias{fm_network_cycle_error}
\title{Compute Directed 3-Cycle Error}
\usage{
fm_network_cycle_error(network, cycle)
}
\arguments{
\item{network}{`fm_network` object.}

\item{cycle}{Character vector of length 3 defining `c(i,j,k)`.}
}
\value{
Frobenius norm cycle error.
}
\description{
Compute Directed 3-Cycle Error
}
