% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fm_network.R
\name{fm_network_cycles}
\alias{fm_network_cycles}
\title{Extract 3-Cycles from Network}
\usage{
fm_network_cycles(network)
}
\arguments{
\item{network}{`fm_network` object.}
}
\value{
List of 3-cycles. For directed networks each cycle is ordered `c(i,j,k)`
  satisfying edges `i->j`, `j->k`, and `k->i`.
}
\description{
Extract 3-Cycles from Network
}
