% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fm_network.R
\name{fm_network_get_map}
\alias{fm_network_get_map}
\title{Get a Map from an FM Network}
\usage{
fm_network_get_map(network, i, j)
}
\arguments{
\item{network}{`fm_network` object.}

\item{i}{Source domain name.}

\item{j}{Target domain name.}
}
\value{
`fm_fit` object or matrix map.
}
\description{
Get a Map from an FM Network
}
