% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fm_diagnostics.R
\name{fm_network_metrics}
\alias{fm_network_metrics}
\title{Compute Network Diagnostic Metrics}
\usage{
fm_network_metrics(network)
}
\arguments{
\item{network}{`fm_network` object.}
}
\value{
List with cycle and transfer consistency diagnostics.
}
\description{
Compute Network Diagnostic Metrics
}
