% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fm_diagnostics.R
\name{fm_network_summary}
\alias{fm_network_summary}
\title{Format Network Report as Text Summary}
\usage{
fm_network_summary(report)
}
\arguments{
\item{report}{`fm_network_report` object.}
}
\value{
Character scalar summary.
}
\description{
Format Network Report as Text Summary
}
