% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fm_package_info.R
\name{fm_package_info}
\alias{fm_package_info}
\title{Package Metadata Helper}
\usage{
fm_package_info()
}
\value{
A named list describing the current package development state.
}
\description{
Package Metadata Helper
}
