% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fm_domain.R
\name{fm_project}
\alias{fm_project}
\title{Project Functions on a Domain Basis}
\usage{
fm_project(domain, x, basis = NULL)
}
\arguments{
\item{domain}{`fm_domain` object.}

\item{x}{Function values on samples.}

\item{basis}{Optional basis list override.}
}
\value{
Basis coefficients.
}
\description{
Project Functions on a Domain Basis
}
