% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fm_utils.R
\name{fm_transfer}
\alias{fm_transfer}
\title{Transfer Functions Through a Functional Map}
\usage{
fm_transfer(fit, x, reverse = FALSE)
}
\arguments{
\item{fit}{`fm_fit` object.}

\item{x}{Function values matrix or vector.}

\item{reverse}{If `TRUE`, transfer target to source.}
}
\value{
Transferred function values on destination domain.
}
\description{
Transfer Functions Through a Functional Map
}
