% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fm_domain.R
\name{fm_unproject}
\alias{fm_unproject}
\title{Reconstruct Functions from Domain Basis Coefficients}
\usage{
fm_unproject(domain, coef, basis = NULL)
}
\arguments{
\item{domain}{`fm_domain` object.}

\item{coef}{Basis coefficients.}

\item{basis}{Optional basis list override.}
}
\value{
Reconstructed function values.
}
\description{
Reconstruct Functions from Domain Basis Coefficients
}
