% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fm_utils.R
\name{is_fm_fit}
\alias{is_fm_fit}
\title{Check Functional Map Fit Class}
\usage{
is_fm_fit(x)
}
\arguments{
\item{x}{Object to check.}
}
\value{
Logical scalar.
}
\description{
Check Functional Map Fit Class
}
